requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# Machine learning and analysis
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Web scraping
selenium>=4.0.0
webdriver-manager>=3.8.0
requests>=2.28.0

# Text processing
textblob>=0.17.0

# Environment and utilities
python-dotenv>=0.19.0

# Optional: For better performance
# numba>=0.56.0
# scipy>=1.9.0
